requirements.txt
fastapi
uvicorn
python-multipart
sqlalchemy
psycopg2-binary
supabase
ollama
google-genai